Add room management and connection handle modules
This commit introduces a new `handle` module for managing WebSocket connection utilities, including methods for sending messages and managing connections. Additionally, a `room` module is added to facilitate room management, allowing clients to join/leave rooms and broadcast messages. The code has been refactored to improve clarity and organization, enhancing the overall structure of the WebSocket server implementation.
